+++ src/trx/game/level/loader.c
@@ -257,17 +257,19 @@ static void M_ReadRoomMesh(
             if (loader->game_version == 1) {
                 vertex->light_base = VFile_ReadS16(file);
                 vertex->is_wibble_disabled = false;
+                vertex->caustics_flags = 0;
                 vertex->color = (RGBA_8888) { 255, 255, 255, 255 };
             } else if (loader->game_version == 2) {
                 VFile_Skip(file, 2);
                 vertex->light_table_value = VFile_ReadU8(file);
                 const uint8_t flags = VFile_ReadU8(file);
                 vertex->is_wibble_disabled = (flags & 0x80u) != 0u;
+                vertex->caustics_flags = 0;
                 vertex->light_base = VFile_ReadS16(file);
                 vertex->color = (RGBA_8888) { 255, 255, 255, 255 };
             } else if (loader->game_version == 3) {
                 VFile_Skip(file, 2); // lighting - unused in TR3
-                VFile_Skip(file, 2); // attributes (water caustics)
+                vertex->caustics_flags = VFile_ReadU16(file); // water caustics
                 vertex->color = M_ARGB1555To8888(VFile_ReadU16(file));
                 vertex->color.a = 255;
                 vertex->light_base = 0;